+++ admin/index.php
@@ -931,7 +931,8 @@ function PAYPAL_adminlist_Category()
     $display = '';
     $sql = "SELECT
                 cat.cat_id, cat.cat_name, cat.description, cat.enabled,
-                cat.grp_access, parent.cat_name as pcat
+                cat.grp_access, parent.cat_name as pcat,
+                cat.discount_pct, cat.discount_beg, cat.discount_end
             FROM {$_TABLES['paypal.categories']} cat
             LEFT JOIN {$_TABLES['paypal.categories']} parent
             ON cat.parent_id = parent.cat_id";
@@ -951,6 +952,8 @@ function PAYPAL_adminlist_Category()
                 'field' => 'description', 'sort' => false),
         array('text' => $LANG_PP['parent_cat'],
                 'field' => 'pcat', 'sort' => true),
+        array('text' => $LANG_PP['discount'],
+                'field' => 'discount_pct', 'sort' => 'false', 'align'=>'right'),
         array('text' => $LANG_PP['visible_to'],
                 'field' => 'grp_access', 'sort' => false),
         array('text' => $LANG_ADMIN['delete'] .
@@ -1005,6 +1008,8 @@ function getAdminField_Category($fieldname, $fieldvalue, $A, $icon_arr)
 
     $retval = '';
     static $grp_names = array();
+    static $now = NULL;
+    if ($now === NULL) $now = PAYPAL_now()->format('Y-m-d');
 
     switch($fieldname) {
     case 'edit':
@@ -1055,6 +1060,14 @@ function getAdminField_Category($fieldname, $fieldvalue, $A, $icon_arr)
         $retval = strip_tags($fieldvalue);
         if (utf8_strlen($retval) > 80) {
             $retval = substr($retval, 0, 80 ) . '...';
+        }
+        break;
+
+    case 'discount_pct':
+        if ($A['discount_end'] < $now || $A['discount_beg'] > $now) {
+            $retval = '';
+        } else {
+            $retval = sprintf('%.03f%%', $fieldvalue);
         }
         break;